+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: demikael <pinheiromikael96@gmail.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/08 21:44:55 by demikael          #+#    #+#             */
-/*   Updated: 2021/09/09 02:26:56 by demikael         ###   ########.fr       */
+/*   Updated: 2021/09/09 15:56:19 by demikael         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 int		ft_printf(const char *, ...);
 
-char	*ft_itoa_base(unsigned long int number, char *base);
+char	*ft_itoa_base(unsigned int number, char *base);
 int		ft_put_char(const char **chr, char ap);
 int		ft_put_decimal(int decimal);
 int		ft_put_hexadecimal(const char **fmt, unsigned long hex);
